chore(shared): fix unnecessary_cast clippy drift in event_setup_rls
Two `.bind(n as i32)` call sites cast from an integer range binding
that already resolves to i32 by inference. Removing the redundant
cast satisfies clippy::unnecessary_cast under -D warnings, which the
PR 2a CI workflow will enforce.

No behavior change; strict-clippy hygiene only.

+++ shared/tests/event_setup_rls.rs
@@ -96,7 +96,7 @@ async fn seed_full_stack(tx: &mut Transaction<'_, Postgres>, name: &str) -> Tena
         )
         .bind(club_id)
         .bind(event_id)
-        .bind(day_number as i32)
+        .bind(day_number)
         // Distinct dates per day, rendered as ISO text so this test
         // file does not need to reach into chrono types directly.
         .bind(format!("2026-11-{:02}", day_number + 10))
@@ -112,7 +112,7 @@ async fn seed_full_stack(tx: &mut Transaction<'_, Postgres>, name: &str) -> Tena
             )
             .bind(club_id)
             .bind(day_id)
-            .bind(trial_number as i32)
+            .bind(trial_number)
             .fetch_one(&mut **tx)
             .await
             .expect("insert trial");
